[UI/Table/Create/240323Part2] - Modify Tables

+++ database/migrations/2024_03_23_074733_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('price');
+            $table->double('price', 8, 2);
             $table->integer('qty_in_stock');
             $table->text('description');
             $table->unsignedBigInteger('status_id');
